[innovation-week-25q1-custom-span] Simplified architecture
